backend/requirements-render.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
python-jose==3.3.0
bcrypt==4.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
mysql-connector-python==9.5.0


